seperated ReviewList component
